svg/geometry/parsing/x-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="8000"><title>SVG Geometry Properties: parsing x with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/geometry.html#X" rel="help"/><h:meta content="x supports the full grammar '&lt;length-percentage&gt;'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("x", "0", "0px");
test_valid_value("x", "-1px");
test_valid_value("x", "calc(2em + 3ex)");
test_valid_value("x", "4%");
test_valid_value("x", "5cm");

  ]]></script></svg>
